Changed Priority to a non-mandatory parameter
In response to discussion around #24, I've updated Priority to be a
non-mandatory parameter for this function.

If you attempt to create a new issue for a project and issue type where
Priority is mandatory, you'll receive an error back from JIRA just as if
you'd omitted a custom field that was mandatory.