Add -ServerName param
This change begins to use the JiraServerConfig functions from the previous commit by allowing the calling function to specify a user-friendly name for a server to the -ServerName parameter of Invoke-JiraMethod. This name will be looked up in the config.json file, and the server's base URL will be prepended to the provided value of the -URI parameter.

If -ServerName is not specified, Invoke-JiraMethod will use the server in the config file marked as default.

This change changes handling of the -URI parameter in Invoke-JiraMethod. Previously, the value of the -URI parameter was expected to be a fully-qualified URI such as `http://example.com/jira/api/2/issue/12345`. After this commit, the value of the -URI parameter should not include the domain name - in that example, the value should be `/jira/api/2/issue/12345`.

This is a breaking change for internal functions, so many tests are failing right now. These will be addressed in upcoming commits.