Add support for full URI again
Some functions, such as Add-JiraIssueComment, use the RestURL property of an output object from another function as a key for their URL. While it is probably possible to rewrite these functions to build a new URL rather than using the existing one, it makes more sense to generalize this case back into Invoke-JiraMethod. That way, this helper method can support both cases.

As a result, this change fixes many of the cases that the previous commit broke. Tests are still failing due to the changes in Get-JiraConfigServer, though.