Complete implementation of -ServerName
All public functions should now have and use the -ServerName parameter, with the exception of the Session functions. Those will need a little bit of extra logic in order to handle authentication correctly.

Use of the functions with basic authentication (-Credential) and the -ServerName parameter should be working correctly at this point.